refactor(broker): move tiny lib.rs utilities into utils/helpers.rs
now_timestamp, format_expiries, and is_dev_mode were three small
pub(crate) fn definitions left at the lib.rs root. Moves them into
utils/helpers.rs (where prune_receipt_claim_journal and the gas
estimators already live) and re-exports them at the crate root so
existing `crate::now_timestamp` / `crate::format_expiries` /
`crate::is_dev_mode` call-sites resolve unchanged.

Drops the now-unused `use std::time::SystemTime;` from lib.rs.
`use boundless_market::contracts::ProofRequest;` stays because
crate::ProofRequest is still reached through the lib.rs `use` import
by storage.rs and db/mod.rs.

lib.rs is now ~120 lines: module declarations, re-exports, the three
batch-related types (Batch / BatchStatus / AggregationState — destined
for aggregator/types.rs in PR-3), and the test / test-utils module
gates.
